update: added features w tick on bottom left side, improved flexbox

+++ client/src/components/Product-Information/Stars.jsx
@@ -14,7 +14,7 @@ const Stars = ({ reviewsList }) => {
       total += currentReview.rating;
     }
     const averageRating = (total / reviewsList.length).toFixed(1);
-    setRating(parseInt(averageRating));
+    setRating(parseFloat(averageRating));
   };
 
   useEffect(() => {
